playwright test

+++ src/auth/auth-provider.tsx
@@ -1,9 +1,12 @@
-import { createContext, useState, useEffect, type ReactNode } from 'react';
-import { decodeToken } from '../utils/helper';
-import { AUTH_ROUTES, PRIVATE_ROUTES } from '../routing/routes';
+import { createContext, useEffect, type ReactNode } from 'react';
 import { useTimeout } from '@mantine/hooks';
 import { notifications } from '@mantine/notifications';
+import { useAppSelector, useAppDispatch } from '../store';
+import { resetUser } from '../store/slices/auth.slice';
+import { getCookie, removeCookie } from '../utils/helper';
+import { AUTH_ROUTES, PRIVATE_ROUTES } from '../routing/routes';
 import { ROLES } from '../utils/constants';
+import { api } from '../api';
 
 const AuthContext = createContext<unknown>(undefined);
 
@@ -12,62 +15,38 @@ interface AuthProviderProps {
 }
 
 const AuthProvider = ({ children }: AuthProviderProps) => {
-  // const [token, setToken, removeToken] = useLocalStorage(LOCAL_STORAGE_KEY, '');
-  // const [token, setToken, removeToken] = useLocalStorage('token', '');
-  const [token, setToken] = useState(localStorage.getItem('token'));
-  // const token = localStorage.getItem('token');
-
-  const [user, setUser] = useState(() => {
-    if (token) {
-      const data: any = decodeToken(token);
-      return data?.user || data;
-    }
-    return {};
-  });
-
-  const role = user?.role || '';
-
-  const isAdmin = role === ROLES.ADMIN;
-
-  // const redirectUrl = role ? REDIRECTION[role] : AUTH_ROUTES.LOGIN.url;
-  const redirectUrl = Object.keys(user).length
-    ? isAdmin
-      ? // ? PRIVATE_ROUTES.ADMIN_DASHBOARD.url
-        PRIVATE_ROUTES.ADMIN_DASHBOARD.url
-      : PRIVATE_ROUTES.DASHBOARD.url
-    : // : AUTH_ROUTES.LOGIN.url;
-      AUTH_ROUTES.LANDING.url;
-
-  const login = (newToken: string) => {
-    localStorage.setItem('token', newToken);
-    setToken(newToken);
-  };
+  const dispatch = useAppDispatch();
+  const { user } = useAppSelector(state => state.auth);
+  const loggedInStatus = getCookie('auth_status');
+  const userRole = getCookie('user_role');
 
   const resetAllStores = () => {};
-
   const { start, clear } = useTimeout(() => resetAllStores(), 1000);
 
-  const logout = () => {
-    // removeToken();
-    localStorage.removeItem('token');
-    localStorage.clear();
-    start();
-    setToken(null);
-    setUser({});
-    notifications.show({
-      message: 'Logged out successfully',
-      color: 'green',
-    });
-  };
+  const logout = async () => {
+    try {
+      const res = await api.auth.logout();
+      if (res.status === 200) {
+        // Clear auth cookies
+        removeCookie('auth_status');
+        removeCookie('user_role');
 
-  useEffect(() => {
-    if (token) {
-      const data: any = decodeToken(token);
-      setUser(data?.user || data);
-    } else {
-      setUser({});
+        dispatch(resetUser());
+        localStorage.clear();
+        start();
+        // notifications.show({
+        //   message: 'Logged out successfully',
+        //   color: 'green',
+        // });
+        return res;
+      }
+    } catch (error) {
+      notifications.show({
+        message: 'Error logging out',
+        color: 'red',
+      });
     }
-  }, [token]);
+  };
 
   useEffect(() => {
     return () => {
@@ -75,15 +54,34 @@ const AuthProvider = ({ children }: AuthProviderProps) => {
     };
   }, [clear]);
 
+  // const redirectUrl = Object.keys(user).length
+  //   ? isAdmin
+  //     ? // ? PRIVATE_ROUTES.ADMIN_DASHBOARD.url
+  //       PRIVATE_ROUTES.ADMIN_DASHBOARD.url
+  //     : PRIVATE_ROUTES.DASHBOARD.url
+  //   : // : AUTH_ROUTES.LOGIN.url;
+  //     AUTH_ROUTES.LANDING.url;
+
+  let redirectUrl = AUTH_ROUTES.LANDING.url;
+  if (loggedInStatus === 'true') {
+    redirectUrl =
+      userRole === ROLES.ADMIN
+        ? PRIVATE_ROUTES.ADMIN_DASHBOARD.url
+        : PRIVATE_ROUTES.DASHBOARD.url;
+  } else if (userRole === ROLES.USER) {
+    redirectUrl = PRIVATE_ROUTES.DASHBOARD.url;
+  } else if (userRole === ROLES.ADMIN) {
+    redirectUrl = PRIVATE_ROUTES.ADMIN_DASHBOARD.url;
+  } else {
+    redirectUrl = AUTH_ROUTES.LANDING.url;
+  }
+
   return (
     <AuthContext.Provider
       value={{
-        // role,
         user,
-        redirectUrl,
-        login,
         logout,
-        // isAdmin,
+        redirectUrl,
       }}
     >
       {children}
